feat: modify code refactoring 1. 로그인 이후 api 접근시 user 중복으로 valid 하는 문제 해결(alarm 목록, principle 로 정보를 넘김) 2. N+1 불필요한 연관관계 쿼리 안나가도록 제어(alarm 목록, feed 목록) 3. Post delete 시 연관된 좋아요, Comment 삭제로직 추가 4. countByPost JPQL -> query method 변경

+++ src/main/java/com/fc/sns/controller/PostController.java
@@ -66,8 +66,8 @@ public class PostController {
     }
 
     @GetMapping("/{postId}/likes")
-    public Response<Integer> likeCount(@PathVariable Long postId) {
-        int count = postService.likeCount(postId);
+    public Response<Long> likeCount(@PathVariable Long postId) {
+        Long count = postService.likeCount(postId);
         return Response.success(count);
     }
     @PostMapping("/{postId}/comments")
